refector: refector file structurer of project
