+++ src/lib/types/storage.ts
@@ -1,3 +1,1 @@
 export type CommandType = { command: string; result: string };
-export type CountType = number;
-export type TestType = string;
