+++ src/routes/+page.svelte
@@ -3,57 +3,13 @@
 	import type { CommandType } from "$lib/types/storage";
 	import { TERMINAL_HISTORY_KEY } from "$lib/constants/localStorageKey";
 	import { getLocalStorageItem } from "$lib/utils/storage";
+	import { UserInputForm } from "./components";
 
-	import {
-		clearStorageArr,
-		findAvailableCommand,
-		historyLengthCutter,
-		isValidCommand,
-		outputCreator,
-		putLocalStorageArr
-	} from "./util";
-	import { WelcomeMessage, HistoryLine, AutoComplete } from "./components";
-	import { COMMANDS } from "./constant";
+	import { historyLengthCutter } from "$lib/utils/command";
+	import { WelcomeMessage, HistoryLine } from "./components";
 
 	let commandArr: CommandType[] | undefined = $state([]);
-	let inputCommand = $state("");
 	let inputBind: HTMLElement | undefined = $state();
-
-	const commandOnSubmit = async (event: SubmitEvent) => {
-		event.preventDefault();
-		if (inputCommand == "clear") {
-			commandArr = [];
-			inputCommand = "";
-			clearStorageArr();
-			return;
-		}
-		const commandObject = {
-			command: inputCommand,
-			result: outputCreator(inputCommand)
-		};
-		putLocalStorageArr(commandObject);
-		if (!commandArr) {
-			commandArr = [commandObject];
-			inputCommand = "";
-			await tick();
-			scroll(0, document.body.scrollHeight);
-			return;
-		}
-		const cutArr = historyLengthCutter(commandArr);
-		commandArr = [...cutArr, commandObject];
-		inputCommand = "";
-		await tick();
-		window.scrollTo(0, document.body.scrollHeight);
-	};
-
-	const changeCommandHandler = (command: string) => {
-		inputCommand = command;
-	};
-
-	let availableCommands = $derived(findAvailableCommand(inputCommand));
-	let isWholeCommand = $derived(
-		Object.values(COMMANDS).find((value) => value === inputCommand)
-	);
 
 	onMount(async () => {
 		let history = getLocalStorageItem(TERMINAL_HISTORY_KEY);
@@ -77,60 +33,12 @@
 				<HistoryLine {command} />
 			{/each}
 		{/if}
-		<form onsubmit={commandOnSubmit} autocomplete="off" class="formContainer">
-			{#if availableCommands.length !== 0 && !isWholeCommand}
-				<AutoComplete
-					onChangeCommand={changeCommandHandler}
-					currentInput={inputCommand}
-					{availableCommands}
-				/>
-			{/if}
-			<p class="userInputCommand">khanne-sh :</p>
-			<input
-				class={`inputTag ${isValidCommand(inputCommand)}`}
-				name="command"
-				bind:value={inputCommand}
-				bind:this={inputBind}
-			/>
-		</form>
+		<UserInputForm bind:commandArr bind:inputBind />
 	</div>
 </main>
 
 <style>
 	.container {
 		padding: 1rem;
-	}
-	.formContainer {
-		display: flex;
-		padding: 0.5rem 0;
-		position: relative;
-	}
-	.userInputCommand {
-		color: #57c6fe;
-		width: 6rem;
-	}
-	.commandOutput {
-		white-space: pre-wrap;
-		color: #f1f0ef;
-	}
-	.previousInput {
-		width: 100%;
-	}
-	.inputTag {
-		width: 100%;
-		border: none;
-		background-color: transparent;
-		caret-color: white;
-		padding: 0;
-		margin-bottom: 2rem;
-	}
-	.inputTag:focus {
-		outline: none;
-	}
-	.validInput {
-		color: #5af68d;
-	}
-	.invalidInput {
-		color: #fe5b56;
 	}
 </style>
