+++ src/routes/components/index.ts
@@ -1,3 +1,3 @@
 export { default as HistoryLine } from "./HistoryLine.svelte";
 export { default as WelcomeMessage } from "./WelcomeMessage.svelte";
-export { default as AutoComplete } from "./AutoComplete/index.svelte";
+export { default as UserInputForm } from "./UserInputForm/index.svelte";
